pyproject.toml
[project]
name = "devguardian"
version = "0.1.0"
description = "An MCP server for AI-powered coding help, debugging, and Git operations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.3.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
]

[project.scripts]
devguardian = "devguardian.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devguardian"]
